Make game adjust to wide screens in a sensible way.
Also... adjust red pulsing animation of timebar, simplify menucards and imagebutton creations by moving defaults into the components.

Ensure that bubbles move to new correct places (and new size) if the game is resized during play, but that the number of rows and columns does not change until a new game is started.